+++ src/pyiem/plot/_mpl.py
@@ -6,11 +6,9 @@ import os
 from io import BytesIO
 
 import geopandas as gpd
-
-# Third party libraries
-import httpx
 import numpy as np
 import rasterio
+import requests
 from PIL import Image
 from pymemcache.client import Client
 from pyproj import Transformer
@@ -69,7 +67,7 @@ def get_tile_data(url):
     res = mc.get(key)
     if res is None:
         LOG.info("Fetching %s", url)
-        resp = httpx.get(url, timeout=10)
+        resp = requests.get(url, timeout=10)
         resp.raise_for_status()
         res = resp.content
         mc.set(key, res)